Add files via upload
We want to build an intelligent system that can calculate the weekly subscription expenses of
essential items for a household. For the sake of this assignment, we will go with the weekly
newspaper subscription. Following table demonstrate prices of some of the prominent
newspapers on a daily basis, all prices in Indian rupees.
Please do not accept the below data from the command line/console, rather use a suitable data
structure to store below values (class/struct).
The input to the program should be the weekly budget/amount that the user has allocated to his
subscriptions. The output must be all possible combinations of the newspaper subscriptions for
the user budget.